container: move frag_vec to container
