+++ src/v/archival/ntp_archiver_service.cc
@@ -31,6 +31,7 @@
 #include "cluster/archival_metadata_stm.h"
 #include "cluster/partition_manager.h"
 #include "config/configuration.h"
+#include "container/fragmented_vector.h"
 #include "features/feature_table.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
@@ -40,7 +41,6 @@
 #include "storage/fs_utils.h"
 #include "storage/ntp_config.h"
 #include "storage/parser.h"
-#include "utils/fragmented_vector.h"
 #include "utils/human.h"
 #include "utils/retry_chain_node.h"
 #include "utils/stream_utils.h"
