+++ src/v/cloud_storage/partition_manifest.h
@@ -12,6 +12,7 @@
 
 #include "cloud_storage/base_manifest.h"
 #include "cloud_storage/types.h"
+#include "container/fragmented_vector.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/record.h"
@@ -19,7 +20,6 @@
 #include "segment_meta_cstore.h"
 #include "serde/envelope.h"
 #include "serde/serde.h"
-#include "utils/fragmented_vector.h"
 #include "utils/tracking_allocator.h"
 
 #include <seastar/core/iostream.hh>
