+++ src/v/cluster/cloud_metadata/offsets_lookup_rpc_types.h
@@ -10,10 +10,10 @@
 #pragma once
 
 #include "cluster/errc.h"
+#include "container/fragmented_vector.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "serde/envelope.h"
-#include "utils/fragmented_vector.h"
 
 namespace cluster::cloud_metadata {
