+++ src/v/cluster/cloud_metadata/offsets_snapshot.h
@@ -11,10 +11,10 @@
 
 #include "base/outcome.h"
 #include "cluster/cloud_metadata/error_outcome.h"
+#include "container/fragmented_vector.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "serde/envelope.h"
-#include "utils/fragmented_vector.h"
 
 namespace cluster::cloud_metadata {
