+++ src/v/cluster/cloud_metadata/offsets_upload_rpc_types.h
@@ -11,11 +11,11 @@
 
 #include "cluster/cloud_metadata/types.h"
 #include "cluster/errc.h"
+#include "container/fragmented_vector.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/timeout_clock.h"
 #include "serde/envelope.h"
-#include "utils/fragmented_vector.h"
 
 namespace cluster::cloud_metadata {
