+++ src/v/cluster/cloud_metadata/uploader.cc
@@ -18,6 +18,7 @@
 #include "cluster/cloud_metadata/offsets_upload_rpc_types.h"
 #include "cluster/logger.h"
 #include "cluster/topic_table.h"
+#include "container/fragmented_vector.h"
 #include "model/fundamental.h"
 #include "model/namespace.h"
 #include "raft/consensus.h"
@@ -27,7 +28,6 @@
 #include "ssx/sleep_abortable.h"
 #include "storage/api.h"
 #include "storage/snapshot.h"
-#include "utils/fragmented_vector.h"
 
 #include <seastar/core/condition-variable.hh>
 #include <seastar/core/coroutine.hh>
