+++ src/v/cluster/rm_stm.h
@@ -18,6 +18,7 @@
 #include "cluster/types.h"
 #include "config/configuration.h"
 #include "config/property.h"
+#include "container/fragmented_vector.h"
 #include "features/feature_table.h"
 #include "metrics/metrics.h"
 #include "model/fundamental.h"
@@ -31,7 +32,6 @@
 #include "storage/snapshot.h"
 #include "utils/available_promise.h"
 #include "utils/expiring_promise.h"
-#include "utils/fragmented_vector.h"
 #include "utils/mutex.h"
 #include "utils/prefix_logger.h"
 #include "utils/tracking_allocator.h"
