+++ src/v/container/CMakeLists.txt
@@ -1,3 +1,5 @@
 v_cc_library(
   NAME container
 )
+
+add_subdirectory(tests)
