+++ src/v/kafka/protocol/offset_fetch.h
@@ -13,12 +13,12 @@
 
 #include "base/seastarx.h"
 #include "bytes/iobuf.h"
+#include "container/fragmented_vector.h"
 #include "kafka/protocol/errors.h"
 #include "kafka/protocol/schemata/offset_fetch_request.h"
 #include "kafka/protocol/schemata/offset_fetch_response.h"
 #include "kafka/types.h"
 #include "model/fundamental.h"
-#include "utils/fragmented_vector.h"
 
 #include <seastar/core/future.hh>
