+++ src/v/kafka/protocol/schemata/generator.py
@@ -1098,7 +1098,7 @@ HEADER_TEMPLATE = """
 #include "model/metadata.h"
 #include "kafka/protocol/errors.h"
 #include "base/seastarx.h"
-#include "utils/fragmented_vector.h"
+#include "container/fragmented_vector.h"
 {%- if not struct.is_streamable %}
 #include <fmt/format.h>
 {%- endif %}
