+++ src/v/kafka/server/usage_aggregator.h
@@ -11,9 +11,9 @@
 
 #pragma once
 #include "base/vlog.h"
+#include "container/fragmented_vector.h"
 #include "kafka/server/logger.h"
 #include "storage/kvstore.h"
-#include "utils/fragmented_vector.h"
 
 #include <seastar/core/gate.hh>
 #include <seastar/core/timer.hh>
