+++ src/v/reflection/type_traits.h
@@ -12,8 +12,8 @@
 #pragma once
 
 #include "base/seastarx.h"
+#include "container/fragmented_vector.h"
 #include "tristate.h"
-#include "utils/fragmented_vector.h"
 #include "utils/named_type.h"
 
 #include <seastar/core/chunked_fifo.hh>
