+++ src/v/storage/types.h
@@ -11,6 +11,7 @@
 
 #pragma once
 
+#include "container/fragmented_vector.h"
 #include "model/fundamental.h"
 #include "model/limits.h"
 #include "model/record.h"
@@ -21,7 +22,6 @@
 #include "storage/key_offset_map.h"
 #include "storage/scoped_file_tracker.h"
 #include "tristate.h"
-#include "utils/fragmented_vector.h"
 
 #include <seastar/core/abort_source.hh>
 #include <seastar/core/file.hh> //io_priority
