+++ src/v/utils/tests/CMakeLists.txt
@@ -19,7 +19,6 @@ rp_test(
     uuid_test.cc
     vint_test.cc
     waiter_queue_test.cc
-    fragmented_vector_async_test.cc
   LIBRARIES v::seastar_testing_main v::utils v::bytes v::version absl::flat_hash_set
   ARGS "-- -c 1"
   LABELS utils
@@ -42,7 +41,6 @@ rp_test(
   SOURCES
     bottomless_token_bucket_test.cc
     filtered_lower_bound_test.cc
-    fragmented_vector_test.cc
     human_test.cc
     move_canary_test.cc
     moving_average_test.cc
